Abstracted Guitar config to Config sub component
